--- src/main/java/com/syslab/backend/domain/user/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.syslab.backend.domain.user.repository;
-
-import com.syslab.backend.domain.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, String> {
-}
