--- src/main/java/com/syslab/backend/global/exception/GlobalExceptionHandler.java
@@ -1,54 +1,0 @@
-package com.syslab.backend.global.exception;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    // 아이디/비밀번호 불일치 → 401
-    @ExceptionHandler(InvalidCredentialsException.class)
-    public ResponseEntity<ErrorResponse> handleInvalidCredentials(
-            InvalidCredentialsException ex, HttpServletRequest request) {
-        ErrorResponse body = new ErrorResponse(
-                "https://diveon.net/problems/invalid-credentials",
-                "Unauthorized",
-                401,
-                ex.getMessage(),
-                request.getRequestURI()
-        );
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(body);
-    }
-
-    // 유저 없음 → 404
-    @ExceptionHandler(UserNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleUserNotFound(
-            UserNotFoundException ex, HttpServletRequest request) {
-        ErrorResponse body = new ErrorResponse(
-                "https://diveon.net/problems/user-not-found",
-                "Not Found",
-                404,
-                ex.getMessage(),
-                request.getRequestURI()
-        );
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(body);
-    }
-
-    // 필수 파라미터 누락 → 400
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ErrorResponse> handleValidation(
-            MethodArgumentNotValidException ex, HttpServletRequest request) {
-        ErrorResponse body = new ErrorResponse(
-                "https://diveon.net/problems/bad-request",
-                "Bad Request",
-                400,
-                "필수 파라미터가 누락되었습니다.",
-                request.getRequestURI()
-        );
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(body);
-    }
-}
