--- src/main/java/com/syslab/backend/global/exception/UserNotFoundException.java
@@ -1,8 +1,0 @@
-package com.syslab.backend.global.exception;
-
-public class UserNotFoundException extends RuntimeException {
-
-    public UserNotFoundException() {
-        super("존재하지 않는 사용자입니다.");
-    }
-}
