--- src/main/java/com/syslab/backend/global/response/ApiResponse.java
@@ -1,22 +1,0 @@
-package com.syslab.backend.global.response;
-
-public class ApiResponse<T> {
-
-    private final int status;
-    private final String message;
-    private final T data;
-
-    private ApiResponse(int status, String message, T data) {
-        this.status = status;
-        this.message = message;
-        this.data = data;
-    }
-
-    public static <T> ApiResponse<T> success(String message, T data) {
-        return new ApiResponse<>(200, message, data);
-    }
-
-    public int getStatus() { return status; }
-    public String getMessage() { return message; }
-    public T getData() { return data; }
-}
