+++ src/main/java/net/diveon/backend/domain/user/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package net.diveon.backend.domain.user.repository;
+
+import net.diveon.backend.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, String> {
+}
