+++ src/main/java/net/diveon/backend/global/exception/UserNotFoundException.java
@@ -1,0 +1,8 @@
+package net.diveon.backend.global.exception;
+
+public class UserNotFoundException extends RuntimeException {
+
+    public UserNotFoundException() {
+        super("존재하지 않는 사용자입니다.");
+    }
+}
