+++ src/main/java/net/diveon/backend/global/response/ApiResponse.java
@@ -1,0 +1,22 @@
+package net.diveon.backend.global.response;
+
+public class ApiResponse<T> {
+
+    private final int status;
+    private final String message;
+    private final T data;
+
+    private ApiResponse(int status, String message, T data) {
+        this.status = status;
+        this.message = message;
+        this.data = data;
+    }
+
+    public static <T> ApiResponse<T> success(String message, T data) {
+        return new ApiResponse<>(200, message, data);
+    }
+
+    public int getStatus() { return status; }
+    public String getMessage() { return message; }
+    public T getData() { return data; }
+}
