+++ src/main/java/net/diveon/backend/global/security/JwtProvider.java
@@ -1,0 +1,71 @@
+package net.diveon.backend.global.security;
+
+import io.jsonwebtoken.JwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+import java.nio.charset.StandardCharsets;
+import java.util.Date;
+
+@Component
+public class JwtProvider {
+
+    private final SecretKey key;
+    private final String issuer;
+    private final long accessTokenValidityMs;
+    private static final long REFRESH_TOKEN_VALIDITY_MS = 7 * 24 * 60 * 60 * 1000L; // 7일
+
+    public JwtProvider(
+            @Value("${auth.jwt.secret}") String secret,
+            @Value("${auth.jwt.issuer}") String issuer,
+            @Value("${auth.jwt.token-validity-in-seconds}") long tokenValiditySeconds) {
+        this.key = Keys.hmacShaKeyFor(secret.getBytes(StandardCharsets.UTF_8));
+        this.issuer = issuer;
+        this.accessTokenValidityMs = tokenValiditySeconds * 1000;
+    }
+
+    // access token 생성 (유효기간: application-local.yml token-validity-in-seconds)
+    public String generateAccessToken(String userId) {
+        Date now = new Date();
+        return Jwts.builder()
+                .subject(userId)
+                .issuer(issuer)
+                .issuedAt(now)
+                .expiration(new Date(now.getTime() + accessTokenValidityMs))
+                .signWith(key)
+                .compact();
+    }
+
+    // refresh token 생성 (유효기간: 7일)
+    public String generateRefreshToken(String userId) {
+        Date now = new Date();
+        return Jwts.builder()
+                .subject(userId)
+                .issuer(issuer)
+                .issuedAt(now)
+                .expiration(new Date(now.getTime() + REFRESH_TOKEN_VALIDITY_MS))
+                .signWith(key)
+                .compact();
+    }
+
+    // 토큰 유효성 검증 (서명 + 만료 체크)
+    public boolean validateToken(String token) {
+        try {
+            Jwts.parser().verifyWith(key).build().parseSignedClaims(token);
+            return true;
+        } catch (JwtException | IllegalArgumentException e) {
+            return false;
+        }
+    }
+
+    // 토큰에서 userId 추출
+    public String getUserId(String token) {
+        return Jwts.parser().verifyWith(key).build()
+                .parseSignedClaims(token)
+                .getPayload()
+                .getSubject();
+    }
+}
