+++ src/test/java/net/diveon/backend/BackendApplicationTests.java
@@ -1,0 +1,13 @@
+package net.diveon.backend;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackendApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
